requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=2.0
web3>=6.0.0
groq>=0.3.0
aiohttp>=3.8.0
eth-account>=0.5.0
python-dateutil>=2.8.0